+++ mobile/app/donate.tsx
@@ -1,0 +1,5 @@
+import { DonateScreen } from '../src/screens/donate/DonateScreen';
+
+export default function DonateRoute() {
+  return <DonateScreen />;
+}